pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "google-generativeai>=0.8.4",
    "pypdf2>=3.0.1",
    "streamlit>=1.43.1",
]
